refactor: Checkbox 컴포넌트 분리
